updation

+++ springboot_automation.py
@@ -2,28 +2,31 @@ import os
 import shutil
 
 
-projectName = input("project name :")
+projectName = input("enter project name ")
 
 # copy files recursivley
 dest="/Users/thej/test-project/"+projectName
 
 # Copy the content of
 # source to destination
-destination = shutil.copytree("/Users/thej/test-project/src", dest)
+destination = shutil.copytree("/Users/thej/test-project/demo", dest)
 print("After copying file:")
 
 #iterate over each file and update the project name
 for dname, dirs, files in os.walk(dest):
     for fname in files:
-        
         fpath = os.path.join(dname, fname)
-        with open(fpath) as f:
-            s = f.read()
-        s = s.replace("${project}", projectName)
+        print(fpath)
+        if(fpath.endswith(".jar") or fpath.endswith(".DS_Store")):
+            continue
+        else:
+            with open(fpath) as f:
+                s = f.read()
+            s = s.replace("${project}", projectName)
 
-        os.remove(fpath)
-        fname = fname.replace("${project}", projectName)
-        fpath = os.path.join(dname, fname)
+            os.remove(fpath)
+            fname = fname.replace("${project}", projectName)
+            fpath = os.path.join(dname, fname)
 
-        with open(fpath, "w") as f:
-            f.write(s)
+            with open(fpath, "w") as f:
+                f.write(s)
